chore: move literals into a separate file
